+++ packages/plugins/typescript/jit-sdk/src/index.ts
@@ -1,6 +1,6 @@
-import { PluginFunction, PluginValidateFn, Types } from '@graphql-codegen/plugin-helpers';
+import { oldVisit, PluginFunction, PluginValidateFn, Types } from '@graphql-codegen/plugin-helpers';
 import { LoadedFragment, RawClientSideBasePluginConfig } from '@graphql-codegen/visitor-plugin-common';
-import { concatAST, FragmentDefinitionNode, GraphQLSchema, Kind, visit } from 'graphql';
+import { concatAST, FragmentDefinitionNode, GraphQLSchema, Kind } from 'graphql';
 import { extname } from 'path';
 import { RawJitSdkPluginConfig } from './config';
 import { JitSdkVisitor } from './visitor';
@@ -27,7 +27,7 @@ export const plugin: PluginFunction<RawJitSdkPluginConfig> = (
     ...(config.externalFragments || []),
   ];
   const visitor = new JitSdkVisitor(schema, allFragments, config);
-  const visitorResult = visit(allAst, { leave: visitor });
+  const visitorResult = oldVisit(allAst, { leave: visitor });
 
   return {
     prepend: visitor.getImports(),