+++ packages/plugins/typescript/vue-apollo-smart-ops/src/index.ts
@@ -1,6 +1,6 @@
 import { extname } from 'path';
-import { Types, PluginValidateFn, PluginFunction } from '@graphql-codegen/plugin-helpers';
-import { visit, GraphQLSchema, concatAST, Kind, FragmentDefinitionNode } from 'graphql';
+import { Types, PluginValidateFn, PluginFunction, oldVisit } from '@graphql-codegen/plugin-helpers';
+import { GraphQLSchema, concatAST, Kind, FragmentDefinitionNode } from 'graphql';
 import { LoadedFragment } from '@graphql-codegen/visitor-plugin-common';
 import { VueApolloVisitor } from './visitor';
 import { VueApolloSmartOpsRawPluginConfig } from './config';
@@ -25,7 +25,7 @@ export const plugin: PluginFunction<VueApolloSmartOpsRawPluginConfig, Types.Comp
   ];
 
   const visitor = new VueApolloVisitor(schema, allFragments, config, documents);
-  const visitorResult = visit(allAst, { leave: visitor });
+  const visitorResult = oldVisit(allAst, { leave: visitor });
 
   return {
     prepend: visitor.getImports(),