+++ packages/presets/graphql-modules/tests/builder.spec.ts
@@ -73,12 +73,46 @@ test('should generate interface field resolvers', () => {
       shouldDeclare: false,
       rootTypes: ROOT_TYPES,
       baseVisitor,
+      useGraphQLModules: true,
     }
   );
 
   expect(output).toContain(`BaseUser: 'id' | 'email';`);
   expect(output).toContain(`export type BaseUser = Pick<core.BaseUser, DefinedFields['BaseUser']>;`);
   expect(output).toContain(`export type BaseUserResolvers = Pick<core.BaseUserResolvers, DefinedFields['BaseUser']>;`);
+});
+
+test('should not generate graphql-modules code when useGraphQLModules=false', () => {
+  const output = buildModule(
+    'test',
+    parse(/* GraphQL */ `
+      interface BaseUser {
+        id: ID!
+        email: String!
+      }
+
+      type User implements BaseUser {
+        id: ID!
+        email: String!
+      }
+
+      type Query {
+        me: BaseUser!
+      }
+    `),
+    {
+      importPath: '../types',
+      importNamespace: 'core',
+      encapsulate: 'none',
+      shouldDeclare: false,
+      rootTypes: ROOT_TYPES,
+      baseVisitor,
+      useGraphQLModules: false,
+    }
+  );
+
+  expect(output).not.toContain(`graphql-modules`);
+  expect(output).not.toContain(`gm.`);
 });
 
 test('should generate interface extensions field resolvers ', () => {
@@ -100,6 +134,7 @@ test('should generate interface extensions field resolvers ', () => {
       shouldDeclare: false,
       rootTypes: ROOT_TYPES,
       baseVisitor,
+      useGraphQLModules: true,
     }
   );
 
@@ -116,6 +151,7 @@ test('should include import statement', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -131,6 +167,7 @@ test('should work with naming conventions', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toContain(`Pick<core.Query_RootResolvers, `);
@@ -145,6 +182,7 @@ test('encapsulate: should wrap correctly with namespace', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`export namespace TestModule {`);
@@ -159,6 +197,7 @@ test('encapsulate: should wrap correctly with a declared namespace', () => {
     shouldDeclare: true,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`declare namespace TestModule {`);
@@ -172,6 +211,7 @@ test('encapsulate: should wrap correctly with prefix', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toMatchSnapshot();
@@ -195,6 +235,7 @@ test('should pick fields from defined and extended types', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -227,6 +268,7 @@ test('should reexport used types but not defined in module', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -245,6 +287,7 @@ test('should export partial types, only those defined in module or root types', 
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -275,6 +318,7 @@ test('should export partial types of scalars, only those defined in module or ro
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -295,6 +339,7 @@ test('should use and export resolver signatures of types defined or extended in 
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -332,6 +377,7 @@ test('should not generate resolver signatures of types that are not defined or e
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).not.toContain('CommentResolvers');
@@ -345,6 +391,7 @@ test('should generate an aggregation of individual resolver signatures', () => {
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toBeSimilarStringTo(`
@@ -366,6 +413,7 @@ test('should generate a signature for ResolveMiddleware (with widlcards)', () =>
     shouldDeclare: false,
     rootTypes: ROOT_TYPES,
     baseVisitor,
+    useGraphQLModules: true,
   });
 
   expect(output).toContain(`import * as gm from "graphql-modules";`);