+++ packages/presets/graphql-modules/tests/integration.spec.ts
@@ -87,6 +87,30 @@ describe('Integration', () => {
     expect(output[4].content).toMatch(importStatement);
   });
 
+  test('should allow to disable graphql-modules', async () => {
+    const output = await executeCodegen({
+      generates: {
+        './tests/test-files/modules': {
+          schema: './tests/test-files/modules/*/types/*.graphql',
+          plugins: ['typescript', 'typescript-resolvers'],
+          preset: 'graphql-modules',
+          presetConfig: {
+            importBaseTypesFrom: '@types',
+            baseTypesPath: 'global-types.ts',
+            filename: 'module-types.ts',
+            encapsulateModuleTypes: 'none',
+            useGraphQLModules: false,
+          },
+        },
+      },
+    });
+
+    for (const record of output) {
+      expect(record).not.toContain(`graphql-modules`);
+      expect(record).not.toContain(`gm.`);
+    }
+  });
+
   test('each module-types should include a relative import to glob-types module', async () => {
     const output = await executeCodegen(options);
     const importStatement = `import * as Types from "../global-types";`;