+++ packages/utils/plugins-helpers/src/index.ts
@@ -5,3 +5,4 @@ export * from './helpers';
 export * from './federation';
 export * from './errors';
 export * from './getCachedDocumentNodeFromSchema';
+export * from './oldVisit';