+++ packages/graphql-codegen-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphql-codegen/cli
 
+## 2.2.2
+
+### Patch Changes
+
+- Updated dependencies [97ddb487a]
+  - @graphql-codegen/core@2.3.0
+  - @graphql-codegen/plugin-helpers@2.3.0
+
 ## 2.2.1
 
 ### Patch Changes