+++ packages/plugins/c-sharp/c-sharp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-codegen/c-sharp
 
+## 4.2.0
+
+### Minor Changes
+
+- 97ddb487a: feat: GraphQL v16 compatibility
+
+### Patch Changes
+
+- Updated dependencies [97ddb487a]
+  - @graphql-codegen/visitor-plugin-common@2.5.0
+  - @graphql-codegen/plugin-helpers@2.3.0
+
 ## 4.1.6
 
 ### Patch Changes