+++ packages/plugins/flow/resolvers/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @graphql-codegen/flow-resolvers
 
+## 2.2.0
+
+### Minor Changes
+
+- 97ddb487a: feat: GraphQL v16 compatibility
+
+### Patch Changes
+
+- Updated dependencies [97ddb487a]
+  - @graphql-codegen/flow@2.2.0
+  - @graphql-codegen/visitor-plugin-common@2.5.0
+  - @graphql-codegen/plugin-helpers@2.3.0
+
 ## 2.1.6
 
 ### Patch Changes