+++ packages/plugins/typescript/graphql-apollo/CHANGELOG.md
@@ -1,6 +1,19 @@
 # @graphql-codegen/typescript-graphql-apollo
 
+## 1.1.0
+
+### Minor Changes
+
+- 97ddb487a: feat: GraphQL v16 compatibility
+
+### Patch Changes
+
+- Updated dependencies [97ddb487a]
+  - @graphql-codegen/visitor-plugin-common@2.5.0
+  - @graphql-codegen/plugin-helpers@2.3.0
+
 ## 1.0.0
+
 ### Major Changes
 
 - 979bdc01e: add a new graphql-apollo plugin